requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
langchain==0.1.0
openai==1.3.0
alembic==1.12.1
python-dateutil==2.8.2
